Avoid warnings
Test that the taxonomy slug is an actual taxonomy before adding the
meta box.
The default code produced a warning:
Warning: Invalid argument supplied for foreach()

This happens when used with an invalid taxonomy, which is default with
the
$custom_tax_mb = new WDS_Taxonomy_Radio('custom-tax-slug')
at the end of the code.
This demo line could be commented out, considering it is useful to keep
the configuration code in functions.php, next to the include of the
class, instead of inside the class file. One doesn't need to read the
entire code, when there is excellent instructions in the readme!