Remove reliance on $this when in ajax callback
`$this` isn't accurate once you've made the AJAX request, so we'll remove references to it and recreate the necessary properties to fulfill the request and successfully create the term for the proper taxonomy using the proper input element. Also return false after adding the term to prevent scrolling to the top of the window.