Updated json response messages on invalid add term
Previously there was a block doing multiple checks to ensure it was ok to add
a new term. If it failed a very generic json error response was returned. This
did not allow the user/other devs to understand why a term was not being added.
This commit breaks the checks apart into their own conditionals and provides the
user with better feedback when something has gone amiss.